+++ tools/xdhq/Atlas/NJS/Atlas.js
@@ -161,3 +161,5 @@ module.exports.createTree = () => require('xmlbuilder').create('XDHTML');
 module.exports.DOM = xdhq.XDH;
 
 module.exports.readAsset = xdhq.readAsset;
+module.exports.modes = modes;
+module.exports.mode = mode;